pyproject.toml
[project]
name = "agent-smith-fastino"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.72.0",
    "faiss-cpu>=1.12.0",
    "google-api-python-client>=2.185.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "httpx>=0.28.1",
    "numpy>=2.3.4",
    "openai>=2.6.1",
    "pandas>=2.3.3",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "rapidfuzz>=3.14.1",
    "rich>=14.2.0",
    "scikit-learn>=1.7.2",
    "sentence-transformers>=5.1.2",
    "structlog>=25.5.0",
    "tenacity>=9.1.2",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
    "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "black>=25.9.0",
    "mypy>=1.18.2",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.2",
]
